+++ deployment/bff/app/v1/models/search.py
@@ -2,8 +2,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 
 from pydantic import BaseModel, Field
 
-from deployment.bff.app.v1.models.helper import BaseRequestModel
-from deployment.bff.app.v1.models.modality import ModalityModel
+from deployment.bff.app.v1.models.shared import BaseRequestModel, ModalityModel
 
 _ProtoValueType = Optional[Union[bool, float, str, list, dict]]
 _StructValueType = Union[
