+++ now/constants.py
@@ -4,7 +4,7 @@ from docarray.typing import Image, Text, Video
 
 from now.utils import BetterEnum
 
-DOCKER_BFF_PLAYGROUND_TAG = '0.0.141-test-include-bff-in-offline-flow'
+DOCKER_BFF_PLAYGROUND_TAG = '0.0.141-refactor-delete-unused-code-1'
 NOW_PREPROCESSOR_VERSION = '0.0.118-feat-integrate-elastic-33'
 NOW_ELASTIC_INDEXER_VERSION = '0.0.141-fix-enable-multi-modal-support-13'
 NOW_AUTOCOMPLETE_VERSION = '0.0.8-feat-integrate-elastic-33'
